# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/account": {
      "get": {
        "operationId": "AccountController_listAccounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Account"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Account"
        ]
      },
      "post": {
        "operationId": "AccountController_createAccounts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Account"
        ]
      }
    },
    "/account/{accountId}": {
      "get": {
        "operationId": "AccountController_getAccount",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Account"
        ]
      }
    },
    "/account/{accountId}/link-user": {
      "post": {
        "operationId": "AccountController_linkUserToAccount",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Account"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mynance API",
    "description": "The Mynance API description",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AccountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "last_accessed": {
            "type": "string"
          },
          "iban": {
            "type": "string"
          },
          "bban": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "institution_id": {
            "type": "string"
          },
          "owner_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "created",
          "last_accessed",
          "iban",
          "bban",
          "status",
          "institution_id",
          "owner_name",
          "name"
        ]
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iban": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "iban"
        ]
      },
      "CreateAccountsDto": {
        "type": "object",
        "properties": {
          "externalRequisitionId": {
            "type": "string"
          },
          "accounts": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "externalRequisitionId",
          "accounts"
        ]
      },
...
    }
  }
}